API

Modules

OneTwoMany.OneTwoMany

Types and constants

Functions and macros

OneTwoMany.first_and_rest

OneTwoMany.first_second_rest

OneTwoMany.getsecond

Documentation

OneTwoMany.OneTwoMany — Module

OneTwoMany

Provides provides utility functions for accessing the first, second and the rest of the elements of a collection.

OneTwoMany.first_and_rest — Function

first_and_rest(x)

Returns a tuple that contains the first and the rest of the elements of x.

x may be an AbstractArray, Tuple, NamedTuple, Pair, AbstractString, or an iterable in general.

OneTwoMany.first_second_rest — Function

first_second_rest(x)

Returns a tuple that contains the first, second, and the rest of the elements of x.

x may be an AbstractArray, Tuple, NamedTuple, Pair, AbstractString, or an iterable in general.

OneTwoMany.getsecond — Function

getsecond(x)

Return the second element of x.

x may be an AbstractArray, Tuple, NamedTuple, Pair, AbstractString, or an iterable in general. ```